Guard created
uh, I don't know how to use this but I'll probably make another branch later to fiddle with it. For now, I'm going to start working on implementing functionality into the actual form itself. Storing values and all. Final goal for this branch will be to upload the completed JSON to a google sheet