tab2b functionality implemented
loads and stores values

the way I'm updating cache.json is by quite literally taking the old one, creating a new javascript object with new properties, and then saving it to the same file name, overwriting it. There's probably an easier way of doing this, but fuck it I'm not trying to optimize this app yet.